chore: add ruff, wire it into CI, fix the issues it found
Bring the codebase to a linted, CI-enforced baseline.

- pyproject: curated ruff config (correctness + modernization + simplification),
  with rationale-documented ignores for rules that fight deliberate patterns
  (naive local-time journals, best-effort try/except on optional network work,
  manual subprocess return-code checks, intentional unicode like N⋅m / ₹).
  ruff added to the dev extra.
- .github/workflows/ci.yml: `ruff check .` now runs before the tests.
- Applied ruff's safe fixes across the tree (unused imports, import ordering,
  literal rewrites) and hand-fixed the real ones: a loop-variable-closure smell
  in newscrawl.parse_feed (find() now takes node) and two dead locals in
  produce.py.

`ruff check .` is clean; 156 tests pass.

Still open (logged in ROADMAP): explicit encoding=utf-8 on all file I/O
(Windows cp1252 / the ₹ bug — mitigated today by PYTHONUTF8), and migrating the
end-of-life google.generativeai to google.genai.

+++ src/carshorts/firstframe.py
@@ -53,7 +53,7 @@ def crop_vertical_content(img):
     w, h = img.size
     if h <= 0 or w / h <= 1.2:
         return img
-    content_w = int(round(h * 9 / 16))
+    content_w = round(h * 9 / 16)
     if content_w <= 0 or content_w >= w:
         return img
     left = (w - content_w) // 2
